DEV: Store the normalized referrer version as a smallint
`normalized_referrer_version` only ever holds a small, monotonically
bumped schema version, so a 2-byte `smallint` is more than enough. Using
it instead of a 4-byte `integer` halves the per-row storage for this
column on the high-volume `browser_pageview_events` table.

+++ db/structure.sql
@@ -1918,7 +1918,7 @@ CREATE TABLE public.browser_pageview_events (
     asn integer,
     score integer,
     normalized_referrer character varying(2000),
-    normalized_referrer_version integer
+    normalized_referrer_version smallint
 );
